Enh #10: Use central package module structure
