+++ lib/facter/jboss_fullconfig.rb
@@ -1,3 +1,3 @@
-require File.expand_path(File.join(File.dirname(__FILE__), '../puppet_x/coi/jboss/facts'))
+require File.expand_path(File.join(File.dirname(__FILE__), '../puppet_x/coi/jboss'))
 
 Puppet_X::Coi::Jboss::Facts.define_fullconfig_fact
