+++ lib/puppet/parser/functions/jboss_type_version.rb
@@ -1,4 +1,4 @@
-require File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss/functions/jboss_type_version')
+require File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss')
 
 # Jboss AS private type version function
 #
