+++ lib/puppet/provider/jboss_confignode/jbosscli.rb
@@ -1,8 +1,12 @@
-require File.expand_path(File.join(File.dirname(File.dirname(__FILE__)), 'jbosscli.rb'))
-require File.expand_path(File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss/provider/confignode.rb'))
+require File.expand_path(File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss'))
 
-Puppet::Type.type(:jboss_confignode).provide(:jbosscli, :parent => Puppet::Provider::Jbosscli) do
+Puppet::Type.type(:jboss_confignode).provide(:jbosscli,
+    :parent => Puppet_X::Coi::Jboss::Provider::AbstractJbossCli) do
+
+  desc 'JBoss CLI configuration node provider'
+
   @clean = false
   @data = nil
+
   include Puppet_X::Coi::Jboss::Provider::ConfigNode
 end
