+++ lib/puppet/provider/jboss_securitydomain/jbosscli.rb
@@ -1,6 +1,6 @@
-require File.expand_path(File.join(File.dirname(File.dirname(__FILE__)), 'jbosscli.rb'))
-require File.expand_path(File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss/provider/securitydomain.rb'))
+require File.expand_path(File.join(File.dirname(__FILE__), '../../../puppet_x/coi/jboss'))
 
-Puppet::Type.type(:jboss_securitydomain).provide(:jbosscli, :parent => Puppet::Provider::Jbosscli) do
+Puppet::Type.type(:jboss_securitydomain).provide(:jbosscli,
+    :parent => Puppet_X::Coi::Jboss::Provider::AbstractJbossCli) do
   include Puppet_X::Coi::Jboss::Provider::SecurityDomain
 end
