+++ lib/puppet_x/coi/jboss/configuration.rb
@@ -1,11 +1,5 @@
-# A puppet x module
-module Puppet_X
-# A COI puppet_x module
-module Coi
-# JBoss module
-module Jboss
 # A class for JBoss configuration
-class Configuration
+class Puppet_X::Coi::Jboss::Configuration
   class << self
 
     @config = nil
@@ -91,8 +85,4 @@ class Configuration
     end
 
   end
-end
-
-end
-end
 end
