+++ spec/unit/facter/jboss_fullconfig_spec.rb
@@ -1,5 +1,4 @@
 require 'spec_helper'
-require 'puppet_x/coi/jboss/configuration'
 
 describe 'Fact jboss_fullconfig', :type => :fact do
